move british airport validation out of validator.js
