+++ src/common/utils/validator.js
@@ -1,6 +1,5 @@
 const countries = require('i18n-iso-countries');
 const moment = require('moment');
-const airportCodes = require('./airport_codes.json');
 const ValidationRule = require('../../common/models/ValidationRule.class');
 const freeCirculationValues = require('../seeddata/egar_craft_eu_free_circulation_options.json');
 const visitReasonValues = require('../seeddata/egar_visit_reason_options.json');
@@ -174,18 +173,6 @@ function currentOrFutureDate(dObj) {
     }
     if (dObj.m == currDate.getMonth() + 1) {
       return dObj.d >= currDate.getDate();
-    }
-  }
-  return false;
-}
-
-function isBritishAirport(airports) {
-  if (airports.includes(null) || airports.includes(undefined)) {
-    return true;
-  } else {
-    const britishAirports = airportCodes.filter(value => value.british).map(item => item.id);
-    if (britishAirports.includes(airports[0]) || britishAirports.includes(airports[1])) {
-      return true;
     }
   }
   return false;
@@ -509,5 +496,4 @@ module.exports = {
   handleResponseError,
   sanitiseValue,
   autoTab,
-  isBritishAirport,
 };
